Require serial operation for the `packer_validate` hook
Adjust the `packer_validate` hook's configuration to set the
`require_serial` to true. This will force the hook to run as a single
process instead of in parallel against matching files. This is
necessary because the `packer init` command is run as part of the
hook's operation the plugin cache it accesses is not safe for
concurrent access. This has resulted in observed errors around
concurrent plugin cache access such as:

Error: Missing plugins

The following plugins are required, but not installed:

* github.com/hashicorp/amazon ~> 1.2

Did you run packer init for this project ?

and

Error: Error discovering plugin github.com/hashicorp/amazon

fork/exec
/home/runner/.config/packer/plugins/github.com/hashicorp/amazon/packer-plugin-amazon_v1.3.7_x5.0_linux_amd64:
text file busy